[SDCP-850] Fix organization tag mapping in Semaphore analyze function
The auto-tagger macro runs the Semaphore analyze function and then attaches the tags to the article unless they exist already.

Problem:
The XML response contained tags with the name "Organization" (spelled with a 'z'), but these were not being correctly added to the organisation array in the response_dict. This was because the add_tag function didn't correctly map "Organization" (with a 'z') to "organisation" (with an 's').

Solution:
Modified the add_tag function to explicitly handle the conversion from "Organization" to "organisation". Now, when the function encounters an "Organization" tag in the XML response, it correctly maps it to the organisation array in the response_dict.

Refs: https://sofab.atlassian.net/browse/SDCP-850